requirements.txt
asgiref==3.8.1
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
mysqlclient==2.2.7
sqlparse==0.5.3
